requirements.txt
selenium~=4.3.0
pytest-parallel==0.1.1
pytest-rerunfailures==10.2
pytest~=7.1.2
pytest-html~=3.2.0
allure-python-commons==2.9.45
allure-pytest==2.9.45
Appium-Python-Client~=2.6.0
webdriver-manager

py~=1.11.0
colorama~=0.4.5
h11~=0.13.0
pip~=22.2.2
wheel~=0.36.2
cryptography~=38.0.1
trio~=0.21.0
outcome~=1.2.0
attrs~=22.1.0
sniffio~=1.3.0
sortedcontainers~=2.4.0
cffi~=1.15.1
idna~=3.4
six~=1.16.0
urllib3~=1.26.12
pluggy~=1.0.0
iniconfig~=1.1.1
certifi~=2022.6.15.2
Pillow~=9.2.0
pyparsing~=3.0.9
pycparser~=2.21
setuptools~=57.0.0
wsproto~=1.2.0
tblib~=1.7.0
requests~=2.28.1
psutil~=5.9.2
pexpect~=4.8.0